updated dto sync with entities

+++ Resourcerer_Sol/src/Resourcerer.DataAccess/Entities/CompositeSoldEvent.cs
@@ -1,7 +1,7 @@
 ﻿namespace Resourcerer.DataAccess.Entities;
 public class CompositeSoldEvent : EntityBase
 {
-    public int UnitsSold { get; set; }
+    public double UnitsSold { get; set; }
     public double PriceByUnit { get; set; }
 
     public Guid CompositeId { get; set; }
