+++ Resourcerer_Sol/src/Resourcerer.DataAccess/Entities/ElementPurchasedEvent.cs
@@ -1,7 +1,7 @@
 ﻿namespace Resourcerer.DataAccess.Entities;
 public class ElementPurchasedEvent : EntityBase
 {
-    public int UnitsBought { get; set; }
+    public double UnitsBought { get; set; }
     public double PriceByUnit { get; set; }
     public UnitOfMeasure? UnitOfMeasure { get; set; }
